Fix parameter type mistmatch when calling seedsInitialize function

+++ src/moria.cpp
@@ -24,7 +24,7 @@ void startMoria(int seed, bool start_new_game, bool use_roguelike_keys) {
     // all made into constants or initialized in `globals.cpp`
 
     // Grab a random seed from the clock
-    seedsInitialize(seed);
+    seedsInitialize(static_cast<uint32_t>(seed));
 
     // Init monster and treasure levels for allocate
     initializeMonsterLevels();
